added functionality that will populate the 5 charity cards at page load, and will repopulate the 5 based on the category id if a category is chosen
