+++ server.js
@@ -25,7 +25,7 @@ app.engine(
 app.set("view engine", "handlebars");
 
 // Routes
-require("./routes/apiRoutes")(app);
+//require("./routes/apiRoutes")(app);
 require("./routes/htmlRoutes")(app);
 
 var syncOptions = { force: false };
@@ -48,4 +48,3 @@ db.sequelize.sync(syncOptions).then(function() {
 });
 
 module.exports = app;
-
